WidgetAlgo : Fix `grab()` for PySide2/Qt5
It was emitting the following error :

```
TypeError: 'PySide2.QtGui.QPixmap.grabWindow' called with wrong argument types:
  PySide2.QtGui.QPixmap.grabWindow(int)
Supported signatures:
  PySide2.QtGui.QPixmap.grabWindow(quintptr, int = 0, int = 0, int = -1, int = -1)
```

Also documented the reason we can't update to use Qt 5's preferred screengrab functions, and must instead live with warning output.

+++ python/GafferUI/WidgetAlgo.py
@@ -63,5 +63,11 @@ def grab( widget, imagePath ) :
 	if imageDir and not os.path.isdir( imageDir ) :
 		os.makedirs( imageDir )
 
-	pixmap = QtGui.QPixmap.grabWindow( widget._qtWidget().winId() )
+	# This emits a deprecation warning in Qt5, but the recommended
+	# `QScreen.grabWindow()` alternative is not available in PySide2,
+	# and the alternative `QWidget.grab()` method does not capture
+	# OpenGL content. We could in theory silence the deprecation warning
+	# by temporarily installing a Qt message handler, but surprise surprise,
+	# we can't do that in PySide2 either.
+	pixmap = QtGui.QPixmap.grabWindow( long( widget._qtWidget().winId() ) )
 	pixmap.save( imagePath )
